Logout Dialog

+++ Bookify.UI/Bookify/src/app/components/auth/auth-navbar/auth-navbar.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { StorageService } from 'src/app/services/Storage.Service/storage.service';
 import { ToastService } from 'src/app/services/Toast.Service/toast.service';
+import { LogoutComponent } from '../../Modals/Confirmation/logout/logout.component';
 
 @Component({
   selector: 'app-auth-navbar',
@@ -10,15 +12,27 @@ import { ToastService } from 'src/app/services/Toast.Service/toast.service';
 })
 export class AuthNavbarComponent implements OnInit {
 
-  constructor(private storageService: StorageService, private toastService: ToastService, private router: Router) { }
+  constructor(private storageService: StorageService, private matDialog: MatDialog, 
+    private toastService: ToastService, private router: Router) { }
 
   ngOnInit(): void {
   }
 
   Logout(){
-    this.storageService.Delete('token');
-    this.toastService.openToast(['Logged out Successfully'], "primary");
-    this.router.navigate(['']);
+    
+    const dialogRef = this.matDialog.open(LogoutComponent, {
+      width: '500px',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      var dialogResult = result;
+
+      if(dialogResult){
+        this.storageService.Delete('token');
+        this.toastService.openToast(['Logged out Successfully'], "primary");
+        this.router.navigate(['']);
+      }
+    });
   }
 
 }
